backend/requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==1.6.5
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
googletrans==2.3.0
idna==2.7
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
notebook==5.7.4
numpy==1.14.5
pandas==0.23.1
pandocfilters==1.4.2
pathlib2==2.3.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
pylint==1.9.3
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-Levenshtein==0.12.0
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
rdflib==4.2.2
regex==2018.8.29
requests==2.19.1
scandir==1.9.0
scikit-learn==0.20.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
SPARQLWrapper==1.6.4
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xmlobjects==0.4
xmltodict==0.11.0
django==1.11.13
psycopg2==2.7.7
djangorestframework==3.9.1
django-cors-headers==2.4.0
pylev